Refactor into single project
