--- api/src/main/resources/application.conf
@@ -1,18 +1,0 @@
-hmda {
-  http {
-    filingHost = "0.0.0.0"
-    filingHost = ${?HMDA_HTTP_HOST}
-    adminHost = "0.0.0.0"
-    adminHost = ${?HMDA_HTTP_ADMIN_HOST}
-    publicHost = "0.0.0.0"
-    publicHost = ${?HMDA_HTTP_PUBLIC_HOST}
-    filingPort = 8080
-    filingPort = ${?HMDA_HTTP_PORT}
-    adminPort = 8081
-    adminPort = ${?HMDA_HTTP_ADMIN_PORT}
-    publicPort = 8082
-    publicPort = ${?HMDA_HTTP_PUBLIC_PORT}
-    timeout = 10
-    timeout = ${?HMDA_HTTP_TIMEOUT}
-  }
-}
