+++ build.sbt
@@ -60,74 +60,6 @@ lazy val hmda = (project in file("."))
     ),
     scalafmtSettings,
     dockerSettings,
-    packageSettings
+    packageSettings,
+    libraryDependencies ++= commonDeps ++ akkaDeps ++ akkaPersistenceDeps ++ akkaHttpDeps ++ circeDeps
   )
-  .dependsOn(cluster)
-  .aggregate(
-    model,
-    httpModel,
-    parser,
-    persistence,
-    api,
-    cluster
-  )
-
-lazy val model = (project in file("model"))
-  .settings(hmdaBuildSettings: _*)
-  .settings(
-    libraryDependencies ++= commonDeps ++ akkaDeps
-  )
-
-lazy val httpModel = (project in file("http-model"))
-  .settings(hmdaBuildSettings: _*)
-  .settings(
-    libraryDependencies ++= commonDeps ++ akkaHttpDeps ++ circeDeps
-  )
-  .dependsOn(model)
-
-lazy val parser = (project in file("parser"))
-  .settings(hmdaBuildSettings: _*)
-  .dependsOn(model)
-
-lazy val persistence = (project in file("persistence"))
-  .settings(hmdaBuildSettings: _*)
-  .settings(
-    libraryDependencies ++= commonDeps ++ akkaDeps ++ akkaPersistenceDeps
-  )
-  .dependsOn(model)
-
-lazy val validation = (project in file("validation"))
-  .settings(hmdaBuildSettings: _*)
-  .settings(
-    libraryDependencies ++= commonDeps ++ akkaDeps
-  )
-  .dependsOn(model)
-
-lazy val query = (project in file("query"))
-  .settings(hmdaBuildSettings: _*)
-  .settings(
-    libraryDependencies ++= commonDeps ++ akkaDeps
-  )
-  .dependsOn(model)
-
-lazy val publication = (project in file("publication"))
-  .settings(hmdaBuildSettings: _*)
-  .settings(
-    libraryDependencies ++= commonDeps ++ akkaDeps
-  )
-  .dependsOn(model)
-
-lazy val api = (project in file("api"))
-  .settings(hmdaBuildSettings: _*)
-  .settings(
-    libraryDependencies ++= commonDeps ++ akkaDeps ++ akkaHttpDeps
-  )
-  .dependsOn(httpModel)
-
-lazy val cluster = (project in file("cluster"))
-  .settings(hmdaBuildSettings: _*)
-  .dependsOn(persistence)
-  .dependsOn(validation)
-  .dependsOn(query)
-  .dependsOn(publication)
-  .dependsOn(api)
