--- cluster/src/main/resources/application.conf
@@ -1,102 +1,0 @@
-akka {
-  loggers = ["akka.event.slf4j.Slf4jLogger"]
-  loglevel = "DEBUG"
-  loglevel = ${?HMDA_LOGLEVEL}
-  logging-filter = "akka.event.slf4j.Slf4jLoggingFilter"
-
-  actor {
-    provider = cluster
-  }
-
-  remote {
-    log-remote-lifecycle-events = error
-
-    netty.tcp {
-      hostname = ${?HOST}
-      port = 0
-      port = ${?APP_PORT}
-
-      bind-hostname = 0.0.0.0
-      bind-port = 0
-    }
-
-  }
-
-  cluster {
-    roles = [persistence, validation, query, publication, health, api]
-    failure-detector {
-      //Increased values for AWS environments
-      threshold = 12
-      acceptable-hearbeat-pause = 10s
-    }
-    metrics.enabled = off
-
-    retry-unsuccessful-join-after = 20s
-
-    auto-down-unreachable-after = off
-
-  }
-
-  extensions = [
-    "akka.cluster.pubsub.DistributedPubSub"
-  ]
-
-
-  discovery.method = akka-dns
-
-  management {
-    http {
-      # The hostname where the HTTP Server for Http Cluster Management will be started.
-      # This defines the interface to use.
-      # InetAddress.getLocalHost.getHostAddress is used not overriden or empty
-      hostname = "0.0.0.0"
-      hostname = ${?HOST}
-      port = 19999
-      port = ${?AKKA_HTTP_MANAGEMENT_PORT}
-      bind-hostname = 0.0.0.0
-      bind-port = 0
-    }
-
-    cluster.bootstrap {
-      # Configuration for the first phase of bootstraping, during which contact points are discovered
-      # using the configured service discovery mechanism (e.g. DNS records).
-      contact-point-discovery {
-        # Define this name to be looked up in service discovery for "neighboring" nodes
-        # If undefined, the name will be extracted from the ActorSystem name
-        service-name = ${?MARATHON_APP_ID}
-
-        # Added as suffix to the service-name to build the effective-service name used in the contact-point service lookups
-        # If undefined, nothing will be appended to the service-name.
-        #
-        # Examples, set this to:
-        # "default.svc.cluster.local" or "my-namespace.svc.cluster.local" for kubernetes clusters.
-        service-namespace = ".marathon.mesos"
-        service-namespace = ${?SERVICE_NAMESPACE}
-        # The effective service name is the exact string that will be used to perform service discovery.
-        #
-        # Set this value to a specific string to override the default behaviour of building the effective name by
-        # concatenating the `service-name` with the optional `service-namespace` (e.g. "name.default").
-        effective-name = ${?SERVICE_NAME}
-      }
-      # Configured how we communicate with the contact point once it is discovered
-      contact-point {
-        # If no port is discovered along with the host/ip of a contact point this port will be used as fallback
-        fallback-port = 19999
-        fallback-port = ${?PORT_19999}
-      }
-    }
-  }
-
-}
-
-hmda {
-  runtime-mode = "dev"
-  runtime-mode = ${?HMDA_RUNTIME_MODE}
-  clusterStatusSchedulerStartupDelay = 500
-  clusterStatusSchedulerInterval = 5
-}
-
-clustering {
-  name = "hmda2"
-  name = ${?HMDA_CLUSTER_NAME}
-}
