--- cluster/src/main/scala/hmda/cluster/HmdaClusterRoles.scala
@@ -1,10 +1,0 @@
-package hmda.cluster
-
-object HmdaClusterRoles {
-  final val persistence = "persistence"
-  final val validation = "validation"
-  final val query = "query"
-  final val publication = "publication"
-  final val health = "health"
-  final val api = "api"
-}
