--- http-model/src/main/scala/hmda/http/common/model/Status.scala
@@ -1,6 +1,0 @@
-package hmda.http.common.model
-
-final case class Status(status: String,
-                        service: String,
-                        time: String,
-                        host: String)
