+++ src/main/scala/hmda/HmdaClusterRoles.scala
@@ -1,0 +1,10 @@
+package hmda
+
+object HmdaClusterRoles {
+  final val persistence = "persistence"
+  final val validation = "validation"
+  final val query = "query"
+  final val publication = "publication"
+  final val health = "health"
+  final val api = "api"
+}
