+++ src/main/scala/hmda/http/model/common/HmdaServiceStatus.scala
@@ -1,0 +1,6 @@
+package hmda.http.model.common
+
+final case class HmdaServiceStatus(status: String,
+                                   service: String,
+                                   time: String,
+                                   host: String)
