+++ src/main/scala/hmda/publication/HmdaPublication.scala
